weather_app/requirements.txt
requests~=2.31.0
pytz~=2024.1
pillow~=10.3.0
geopy~=2.4.1
timezonefinder~=6.5.0